+++ CHANGELOG.md
@@ -17,11 +17,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   constructed using `#af_range` can contain.
 - Add the `#af_zip` parser function.
 - Add the `!` overloaded index to show a value.
+- Add the `simple` format.
+- Add the `table` format.
 - Add the `Pages with ArrayFunctions errors` (`af-error-category`) tracking category.
 
 ### Changed
 
 - The `#af_sort` and `#af_ksort` parser functions now compare items normally instead of as strings.
+- Allow `#af_show` to take a format.
+- Localisation updates courtesy of [translatewiki.net](https://translatewiki.net).
+
+### Fixed
+
+- Fix exception when passing the empty string to a keyword argument.
 
 ## [v1.15.0] - 2025-05-13